requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pandas
numpy
pydantic
